classes for extractors
