+++ lib/cuke_slicer.rb
@@ -6,7 +6,8 @@ require "cuke_slicer/filter_set"
 require "cuke_slicer/collections/tag_collection"
 require "cuke_slicer/collections/nested_tag_collection"
 require "cuke_slicer/collections/path_collection"
-require "cuke_slicer/directory_extractor"
+require "cuke_slicer/extractors/directory_extractor"
+require "cuke_slicer/extractors/file_extractor"
 
 
 # Top level module under which the gem code lives.
